toasty/cli.py: add wwtl-sample-image-tiles command
This adds a command that implements a workflow motivated by David Weigel's
usage. If you have a very large sky image without astrometry, you can open it
up in WWT and manually position it. But how do you export it for viewing on
the web?

Our solution here is to ask the user to save their image layer in a WWTL layer
export file containing just the image set layer. This tool then reads in that
file, extracts the underlying (large) image, tiles it, and also translates the
astrometric information appropriately into a WTML file expressing the tiled
output.

The output WTML file still needs editing to absolute-ize the URLs, but it's
pretty close to a one-stop-shop for the data processing.