toasty/image.py: add a `black_to_transparent` image loading option
This option does what it says, setting the alpha channel of all-black
pixels to be fully transparent. Implementation is a bit annoying because
there seems to be no efficient way to modify a PIL image in place, which
is worrisome when we aspire to process gigapixel-class images all in
memory if possible. I tried something that is hopefully efficient, and
we'll see if further refinements are needed.