toasty/samplers.py: round pixel coordinates in plate carree samplers
We were calculating pixel indices as floats and then turning them into
integers with `.astype(np.int)`. This process has truncating behavior,
however, which so on average we end up with a half-pixel shift away from
the ideal pixel choice (since we're using a "nearest" sampling
approach). Start rounding the coordinates, so that index `50.9` starts
landing on pixel 51, not pixel 50.