Update for better parity handling in wwt_data_formats 0.10
In wwt_data_formats 0.10 we'll fix up our WCS processing to correctly
deal with different image parities. This exposes the fact that AstroPix
and Djangoplicity are returning data that, if interpreted directly into
WCS, are a bit wrong. Here we account for that. More work will need to
be done to tighten up the parity handling in toasty overall.