toasty/multi_tan.py: completely revamp
Haven't touched this code in a while, but WWT can now display tiled FITS
so we're about to use it a lot more! The new code has to be careful
about the parities of FITS vs. tilings but I believe that we're doing
that all correctly now. We now also emit WTML using the standard
framework and otherwise modernize this code.

No parallel implementation yet, just trying to tidy up the serial
approach first.