toasty/builder.py: provide better study tiling API
It's helpful to split the study tiling operation into two steps,
preparation and execution, because (1) you can only correctly apply WCS
information after preparation and (2) it's nice to check that the WCS is
all set up correctly before spending the time on the execution step. Here
we add some new APIs to Builder and use them in tile-study

We also improve tile-study to ensure that its image has negative
(JPEG-like) parity, which is required for WWT studies.