toasty/image.py: add a U8 mode
This is for the Kaguya lunar dataset. We're a bit sloppy here by using 0 as the "mask"
value, but it's a reasonable approach. This patch also includes a few fixups to handle
the F64 mode in the same way as F32, where it wasn't added to all of the logic.