toasty/image.py: fix issue updating PIL-based images
We weren't correctly updating Images when they were originally read in
by PIL, because we'd write out the old PIL object and not the updated
array. I thought that the fact that we would get a read-only Numpy array
would prevent such issues, but apparently not? We may need a more
generic API to ensure that such bugs don't reoccur, but there are some
specific cases where it will always be correct to clear the PIL data.