toasty/samplers.py: use generic Image API for the chunked updater
My initial code was hardcoded for the U8 image format, but we should be
better than that. Unfortunately I haven't tested whether it still works
for U8 images.

This is a signpost that the TOAST sampling code should be migrated to
use an Image buffer in general, but I won't do that here.