pipeline: re-fix parity handling in Astropix and Djangoplicity
The bug fixed in WorldWideTelescope/wwt_data_formats#40 was masking a
separate bug in our code here, which in both cases basically converts
AVM metadata to WCS. In #59 I figured out the way to do this right: just
process the AVM and then apply the standard WCS parity-flip operation.
With the wwt_data_formats bug, my old strategy of negating CDELT2 was
precisely equivalent to this, but now that the bug is fixed, we need to
correct the processing here too.

Closes #65.