toasty/collection.py: be more forgiving with FITS shapes
Fix the behavior when the WCS has more axes than the FITS, under the
assumption that these are virtual axes that should be taken to have size
1. Also make it a warning, not a fatal error, to have extra
non-celestial axes with sizes other than 1. It means that we're dropping
data without (currently) a way to do anything about it, but that's more
useful than rejecting the file altogether.