Cythonize the `latlon_tile_filter` function
In processing a FITS-to-TOAST dataset going down to level 14, with about
24 million tiles to process, I discovered that the tile filter function
was a non-trivial bottleneck -- it took about 10 minutes just to count
the tiles that needed to be processed! Some profiling showed that most
of the time was actually in the Numpy min/max functions. Here we
Cythonize the function and avoid the Numpy functions. My measurements
indicate a speedup of a factor of around 20.